Migrate Subjects/Datasets/Gates/Retrieval services (T7.8-T7.11)
WS-M3.2 service migrations completing the per-module relocation track.
All four are mechanical migrations — the M1+ services already used
self._client._request(method, path, json=body), so each becomes:

- BaseService inheritance (self.client.* call shape)
- Imports updated to ..schemas.federation + ..exceptions
- Legacy kamiwaza/<name>.py becomes a thin re-export

Per-task scope:

- T7.8 (ENG-5042) SubjectsAPI + SubjectGrantsAPI →
  kamiwaza_sdk/services/subjects.py. PUT-only upsert per OQ-11
  preserved. lazy-property grants() pattern preserved.

- T7.9 (ENG-5043) DatasetsAPI →
  kamiwaza_sdk/services/datasets.py. Includes set_gate / get_gate /
  clear_gate. Module-name disambiguation: legacy
  kamiwaza_sdk/services/catalog.py covers the broader catalog surface;
  this is the minimal M3 slice setup.py + cmd_m3 reach for.

- T7.10 (ENG-5044) GatesAPI →
  kamiwaza_sdk/services/gates.py. discover() only — full surface
  (set_gate / clear_gate / packages.*) is WS-M5.

- T7.11 (ENG-5045) RetrievalAPI →
  kamiwaza_sdk/services/retrieval_federation.py. Module-name
  disambiguation per design §6.2 T7.11: the legacy
  kamiwaza_sdk.services.retrieval is the broader streaming
  RetrievalService surface (transports, embeddings); this is the M3
  list+cancel slice. Investigated whether to merge — opted to keep
  separate because the two surfaces have different return shapes and
  different callers; future v2 could consolidate.

Full unit suite regression: 1571 pass, 1 unrelated BSD-sed skip
(unchanged from T7.7).

References: ENG-5042/5043/5044/5045, design v0.3.7 §4.2.11 §6.2 T7.8-T7.11.
